requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2   

# Validation & config
pydantic[email]==2.7.1
pydantic-settings==2.3.0

# WebSockets
websockets==12.0

# Utilities
python-multipart==0.0.9
httpx==0.27.0